remove suspense/lazy wrapper components
